# apps/web/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG as a Service API",
    "version": "0.1.0",
    "description": "Multi-tenant retrieval-augmented generation platform. Upload documents into collections, then ask grounded questions and get answers with citations. Authenticate with an API key (`Authorization: Bearer rag_live_…`) for programmatic access, or a Clerk session for dashboard-only endpoints.",
    "contact": {
      "name": "RAG as a Service",
      "url": "https://github.com/your-org/rag-as-a-service"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "rag_live_...",
        "description": "Programmatic API key. Send `Authorization: Bearer rag_live_…` (or the `X-API-Key` header). Rate-limited per key; responses carry RateLimit-* headers."
      },
      "SessionAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session JWT from the dashboard SPA. Required for API-key management and analytics endpoints, which never accept an API key."
      }
    },
    "schemas": {
...
      "AuthContext": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "example": "org_2abc123"
          },
          "authType": {
            "$ref": "#/components/schemas/AuthType"
          },
          "userId": {
            "type": "string",
            "description": "Clerk user id — present only for `session` auth."
          },
          "keyId": {
            "type": "string",
            "description": "Id of the API key used — present only for `apikey` auth."
          }
        },
        "required": [
          "tenantId",
          "authType"
        ]
      },
      "AuthType": {
        "type": "string",
        "enum": [
          "session",
          "apikey"
        ],
        "example": "apikey"
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Collection not found"
          }
        },
        "required": [
          "error"
        ],
        "description": "Standard error response body used across the API."
      },
      "RateLimitErrorBody": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Rate limit exceeded"
          },
          "retryAfter": {
            "type": "integer",
            "example": 30,
            "description": "Seconds until the caller may retry (mirrors Retry-After)."
          },
          "limit": {
            "type": "integer",
            "example": 60,
            "description": "The limit that was exceeded (requests per minute)."
          }
        },
        "required": [
          "error",
          "retryAfter",
          "limit"
        ]
      },
      "CollectionResponse": {
        "type": "object",
        "properties": {
          "collection": {
            "$ref": "#/components/schemas/Collection"
          }
        },
        "required": [
          "collection"
        ]
      },
      "Collection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "col_9f8b2a1c"
          },
          "tenantId": {
            "type": "string",
            "example": "org_2abc123"
          },
          "name": {
            "type": "string",
            "example": "Product docs"
          },
          "description": {
            "type": "string",
            "example": "Public product documentation and FAQs."
          },
          "createdAt": {
            "type": "integer",
            "example": 1753800000000
          },
          "updatedAt": {
            "type": "integer",
            "example": 1753800000000
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "createdAt",
          "updatedAt"
        ],
        "description": "A named group of documents indexed and queried together."
      },
      "CreateCollectionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Product docs"
          },
          "description": {
            "type": "string",
            "example": "Public product documentation and FAQs."
          }
        },
        "required": [
          "name"
        ]
      },
      "ListCollectionsResponse": {
        "type": "object",
        "properties": {
          "collections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Collection"
            }
          }
        },
        "required": [
          "collections"
        ]
      },
...
      "ApiKeyCreateResponse": {
        "type": "object",
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/ApiKey"
          },
          "key": {
            "type": "string",
            "example": "rag_live_a1b2c3d4e5f6g7h8i9j0",
            "description": "Full plaintext key. Never returned again."
          }
        },
        "required": [
          "apiKey",
          "key"
        ]
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "key_1a2b3c4d"
          },
          "name": {
            "type": "string",
            "example": "Production server"
          },
          "keyPrefix": {
            "type": "string",
            "example": "rag_live_a1b2",
            "description": "Leading display portion; not usable to authenticate."
          },
          "last4": {
            "type": "string",
            "example": "9f3c"
          },
          "rateLimitPerMinute": {
            "type": "integer",
            "example": 60
          },
          "createdAt": {
            "type": "integer",
            "example": 1753800000000
          },
          "lastUsedAt": {
            "type": "integer",
            "example": 1753800000000,
            "description": "Last successful auth with this key; null/absent if never used."
          },
          "revokedAt": {
            "type": "integer",
            "example": 1753800000000,
            "description": "When the key was revoked; null/absent while active."
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "last4",
          "rateLimitPerMinute",
          "createdAt"
        ],
        "description": "An API key (never includes key material)."
      },
      "CreateApiKeyRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Production server"
          },
          "rateLimitPerMinute": {
            "type": "integer",
            "example": 120,
            "description": "Optional per-key override; defaults to the server's configured limit."
          }
        },
        "required": [
          "name"
        ]
      },
      "ListApiKeysResponse": {
        "type": "object",
        "properties": {
          "apiKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKey"
            }
          }
        },
        "required": [
          "apiKeys"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/me": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Current authenticated principal",
        "description": "Returns the tenant + auth type resolved from the presented credential.",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "SessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authenticated principal.",
            "headers": {
              "RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per minute for the presented API key."
              },
              "RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current sliding window."
              },
              "RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until the window frees capacity."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContext"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — too many requests for this API key.",
            "headers": {
              "RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per minute for the presented API key."
              },
              "RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current sliding window."
              },
              "RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until the window frees capacity."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds the caller should wait before retrying."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/collections": {
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Create a collection",
        "description": "Create a new, empty collection scoped to the caller's tenant.",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "SessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collection created.",
            "headers": {
              "RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per minute for the presented API key."
              },
              "RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current sliding window."
              },
              "RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until the window frees capacity."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request — malformed body or failed a field constraint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — too many requests for this API key.",
            "headers": {
              "RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per minute for the presented API key."
              },
              "RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current sliding window."
              },
              "RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until the window frees capacity."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds the caller should wait before retrying."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "List collections",
        "description": "List every collection owned by the caller's tenant, newest first.",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "SessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The tenant's collections.",
            "headers": {
              "RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per minute for the presented API key."
              },
              "RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current sliding window."
              },
              "RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until the window frees capacity."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCollectionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — too many requests for this API key.",
            "headers": {
              "RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per minute for the presented API key."
              },
              "RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current sliding window."
              },
              "RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until the window frees capacity."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds the caller should wait before retrying."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/api-keys": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create an API key",
        "description": "Mint a new API key. The plaintext `key` is returned exactly once in this response and never again — store it securely. Session-only: an API key cannot create keys.",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Key created; `key` shown once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request — malformed body or failed a field constraint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing session, or an API key was presented (keys cannot manage keys).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys",
        "description": "List the tenant's API keys. Never returns key material.",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The tenant's API keys.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListApiKeysResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}